Move UserMenu to zetkin-common
