--- src/components/header/UserMenu.scss
@@ -1,76 +1,0 @@
-.UserMenu {
-    position: relative;
-    float: right;
-    color: $c-brand-main;
-}
-.UserMenu-user {
-    position: relative;
-    border-radius: 3em;
-    color: $c-brand-main;
-    cursor: pointer;
-    z-index: 2;
-
-    &:hover {
-        background-color: $c-ui-bg;
-    }
-}
-.UserMenu-expanded .UserMenu-user {
-    background-color: $c-ui-bg;
-}
-.UserMenu-avatar {
-    display: inline-block;
-    margin-right: .5em;
-    height: 3em;
-    width: 3em;
-    border-radius: 50%;
-    background-color: $c-brand-main;
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: 100%;
-    vertical-align: middle;
-}
-.UserMenu-name {
-    display: inline;
-    vertical-align: middle;
-
-    &::after{
-        @include icon("\f0d7");
-        color: $c-ui-dark; 
-        transition: transform 0.16667s ease-out;
-        transform-origin: center 45%;
-    }
-}
-.UserMenu-expanded .UserMenu-name {
-    &::after{
-        transform: rotateZ(180deg)
-    }
-}
-.UserMenu-dropDown {
-    position: absolute;
-    width: 100%;
-    left: 0;
-    top: 50%;
-    margin: 0;
-    padding: 2em 0.5em 0.5em;
-    border-radius: 2px;
-    z-index: 1;
-    box-shadow: 0 1px 2px rgba(0, 0, 0, 0.15);
-    color: $c-text;
-    list-style: none;
-    background-color: $c-text-contrast;
-
-    a {
-        color: inherit;
-        text-decoration: none;
-        &:hover {
-            color: $c-brand-main;
-        }
-    }
-}
-
-.UserMenu-logout:before {
-    @include icon($fa-var-power-off);
-}
-.UserMenu-settings:before {
-    @include icon($fa-var-user);
-}
